+++ packages/vscode-graphql/src/server/index.ts
@@ -1,3 +1,6 @@
+// this lives in the same monorepo! most errors you see in
+// vscode that aren't highlighting or bracket completion
+// related are coming from our LSP server
 import { startServer } from 'graphql-language-service-server';
 
 // The npm scripts are configured to only build this once before